refactor: move problemDiagramId to diagram.ts
avoids cyclic dependency edge.ts <-> store.ts
